Create fibo.py
In fibo.py implemented  Fibonacci numbers.

In mathematics, the Fibonacci numbers, commonly denoted Fn , form a sequence, the Fibonacci sequence,
in which each number is the sum of the two preceding ones.
The sequence commonly starts from 0 and 1, although some authors omit the initial terms and start the sequence from 1 and 1 or from 1 and 2.
Starting from 0 and 1, the first few values in the sequence are:[1]

0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144.